# educa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Educa LRS REST API Backend",
        "description": "Educa LRS API Backend",
        "contact": {
            "email": "support@digitallearning.gmbh"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/actors": {
            "get": {
                "tags": [
                    "Actors"
                ],
                "summary": "List all actors",
                "operationId": "a5f35eaaee011abf3ddc6d8db52f7125",
                "responses": {
                    "200": {
                        "description": "List of actors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "mbox": {
                                                "type": "string",
                                                "example": "mailto:john.doe@example.com"
                                            },
                                            "account_homepage": {
                                                "type": "string",
                                                "example": "http://example.com"
                                            },
                                            "account_name": {
                                                "type": "string",
                                                "example": "johndoe"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Actors"
                ],
                "summary": "Create a new actor",
                "operationId": "1fa2d7dc5ea77254071e874854ce7638",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "mbox"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "mbox": {
                                        "type": "string",
                                        "example": "mailto:john.doe@example.com"
                                    },
                                    "account_homepage": {
                                        "type": "string",
                                        "example": "http://example.com"
                                    },
                                    "account_name": {
                                        "type": "string",
                                        "example": "johndoe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actor created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "mbox": {
                                            "type": "string",
                                            "example": "mailto:john.doe@example.com"
                                        },
                                        "account_homepage": {
                                            "type": "string",
                                            "example": "http://example.com"
                                        },
                                        "account_name": {
                                            "type": "string",
                                            "example": "johndoe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/actors/{id}": {
            "get": {
                "tags": [
                    "Actors"
                ],
                "summary": "Retrieve an actor by ID",
                "operationId": "1b538da5e4c57d7e61d6dea6e271ac35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the actor to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actor retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "mbox": {
                                            "type": "string",
                                            "example": "mailto:john.doe@example.com"
                                        },
                                        "account_homepage": {
                                            "type": "string",
                                            "example": "http://example.com"
                                        },
                                        "account_name": {
                                            "type": "string",
                                            "example": "johndoe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Actors"
                ],
                "summary": "Update actor details",
                "operationId": "64d8d68e06afdc9ee910c6a3a27891f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the actor to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Updated"
                                    },
                                    "mbox": {
                                        "type": "string",
                                        "example": "mailto:updated.john@example.com"
                                    },
                                    "account_homepage": {
                                        "type": "string",
                                        "example": "http://updated.example.com"
                                    },
                                    "account_name": {
                                        "type": "string",
                                        "example": "updatedjohn"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actor updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Updated"
                                        },
                                        "mbox": {
                                            "type": "string",
                                            "example": "mailto:updated.john@example.com"
                                        },
                                        "account_homepage": {
                                            "type": "string",
                                            "example": "http://updated.example.com"
                                        },
                                        "account_name": {
                                            "type": "string",
                                            "example": "updatedjohn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actors"
                ],
                "summary": "Delete an actor",
                "operationId": "5652d042a89f7196e181384312d18ad8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the actor to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actor deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Actor deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}